[fix] integer overflow in hash_old (#218)
introduced in https://github.com/jruby/jruby-openssl/pull/216

+++ src/main/java/org/jruby/ext/openssl/X509Name.java
@@ -633,7 +633,7 @@ public class X509Name extends RubyObject {
 
     @JRubyMethod
     public RubyFixnum hash_old() {
-        int hash;
+        long hash;
         try {
             hash = Name.hashOld( getX500Name() );
         }
